deepagent/requirements.txt
# Core DeepAgent Dependencies
deepagents>=0.1.0
langchain>=0.3.0
langgraph>=0.4.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0

# Search Tools
tavily-python>=0.5.0

# Web UI
fastapi>=0.115.0
uvicorn>=0.34.0
jinja2>=3.1.0
python-multipart>=0.0.18

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
httpx>=0.28.0
aiofiles>=24.0.0

# Optional: For enhanced features
tiktoken>=0.8.0
